+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next release
 
+- Added `\Wizaplace\SDK\Company\CompanyService::updateFile`
+- Added `\Wizaplace\SDK\Company\CompanyService::deleteFile`
+
 ### 1.33.0
 
  - Added `\Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService::getMultiVendorProductById`